refactor(todo): rename TodoModuleConfig to TodoConfig
- Updated references in settings and infrastructure layers.
- Simplified configuration structure for better clarity.
- Enhanced validation and default values in TodoConfig.
