+++ src/moro/config/settings.py
@@ -17,7 +17,7 @@ from pydantic import BaseModel, Field
 
 from moro.modules.common import CommonConfig
 from moro.modules.fantia.config import FantiaConfig
-from moro.modules.todo.config import TodoModuleConfig
+from moro.modules.todo.config import TodoConfig
 
 logger = logging.getLogger(__name__)
 
@@ -39,7 +39,7 @@ class ConfigRepository(BaseModel):
 
     common: CommonConfig = Field(default_factory=CommonConfig)  # Common configuration instance
     fantia: FantiaConfig = Field(default_factory=FantiaConfig)  # Fantia-specific configuration
-    todo: TodoModuleConfig = Field(default_factory=TodoModuleConfig)  # Todo module configuration
+    todo: TodoConfig = Field(default_factory=TodoConfig)  # Todo module configuration
 
     @classmethod
     def create(
@@ -68,10 +68,7 @@ class ConfigRepository(BaseModel):
             binder.bind(ConfigRepository, to=self)
             binder.bind(CommonConfig, to=self.common)
             binder.bind(FantiaConfig, to=self.fantia)
-            binder.bind(TodoModuleConfig, to=self.todo)
-            # Todo module specific config binding
-            from moro.modules.todo.infrastructure import TodoConfig
-            binder.bind(TodoConfig, to=self.todo.todo)
+            binder.bind(TodoConfig, to=self.todo)
 
         return configure
